requirements.txt
# Core LangChain/LangGraph
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0
langchain-groq>=0.1.0
langgraph>=0.0.40

# Vector Database
faiss-cpu>=1.7.4

# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0
openpyxl>=3.1.0

# Image OCR (optional — install manually if needed)
# pytesseract>=0.3.10
# Pillow>=10.0.0

# Embeddings
sentence-transformers>=2.3.0

# Web Framework
streamlit>=1.29.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
typing-extensions>=4.8.0

# Market Data APIs
yfinance>=0.2.32
requests>=2.31.0

# Math/Calculations
numpy>=1.26.0
pandas>=2.1.0

# Caching
cachetools>=5.3.0
